pyproject.toml
[project]
name = "zed-ocr-extractor"
version = "0.1.0"
description = "OCR number extractor for videos/PDFs with API/CLI"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=2.0.0",
    "pillow>=9.0.0",
]

[project.scripts]
ocr-process = "cli.main:cli_entry"
start-server = "server.app:run_server"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-dir]
"" = "."

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
